@@ -76,6 +76,7 @@
     "gulp": "^4.0.2",
     "gulp-concat": "2.6.1",
     "gulp-filter": "5.1.0",
+    "gulp-git": "2.10.1",
     "gulp-header": "2.0.5",
     "gulp-less": "3.5.0",
     "gulp-rename": "^1.4.0",
@@ -83,6 +84,7 @@
     "gulp-sourcemaps": "2.6.4",
     "gulp-uglify": "2.1.0",
     "gulp-util": "3.0.8",
+    "http-server": "^0.12.3",
     "jasmine-check": "0.1.5",
     "jest": "23.6.0",
     "marked": "0.3.19",
@@ -90,9 +92,11 @@
     "mkdirp": "0.5.1",
     "npm-run-all": "4.1.5",
     "prettier": "1.14.2",
-    "react": "^0.12.0",
-    "react-router": "^0.11.2",
+    "react": "17.0.1",
+    "react-dom": "17.0.1",
+    "react-router": "5.2.0",
     "react-tools": "0.13.3",
+    "react-transition-group": "4.4.1",
     "rimraf": "2.6.2",
     "rollup": "0.59.1",
     "rollup-plugin-buble": "0.19.2",
